Add group controller, almost working

+++ TimeKeep/src/main/java/sir/timekeep/rest/GroupController.java
@@ -16,6 +16,8 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
+@RestController
+@RequestMapping("/rest")
 public class GroupController {
     private static final Logger LOG = LoggerFactory.getLogger(GroupController.class);
     private final GroupService groupService;
@@ -26,65 +28,67 @@ public class GroupController {
         this.userService = userService;
     }
 
-    // TO DO: Authorization based on user login
+    @PreAuthorize("hasRole('PREMIUM') && (#id == authentication.principal.user.id)")
     @GetMapping(value ="/{id}/groups", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<List<Group>> findByCreator(@PathVariable Integer id){
         Optional<List<Group>> groups = groupService.findAllByCreator(id);
         return groups.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    // TO DO: Authorization based on user login implementation
-    @PreAuthorize("hasRole('ROLE_PREMIMUM')")
+    @PreAuthorize("hasRole('PREMIUM') && (#id == authentication.principal.user.id)")
     @PostMapping(value = "/{id}/groups", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> createGroup(@PathVariable Integer id, @RequestBody Group group){
         Objects.requireNonNull(group);
-        userService.createGroup(userService.find(id), group);
+        userService.createGroup(userService.find(id).get(), group);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    // TO DO: Authorization based on user login implementation
-    @PreAuthorize("hasRole('ROLE_PREMIMUM') && user_id == groupService.find(group_id).get().getGroupCreator()")
-    @PutMapping(value = "/{id}/groups/{id_group}/add_user", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('PREMIUM') && (#user_id == authentication.principal.user.id)")
+    @PutMapping(value = "/{user_id}/groups/{group_id}/add_user", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addUserToGroup(@PathVariable Integer user_id, @PathVariable Integer group_id, @RequestBody User user){
         Objects.requireNonNull(user);
         if (!groupService.exists(group_id) || userService.find(user_id).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        userService.addUserToGroup(userService.find(user_id), groupService.find(group_id), user);
+        if (!Objects.equals(user_id, groupService.find(group_id).get().getGroupCreator().getId())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        if (groupService.find(group_id).get().getUsers().contains(user)){
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
+        userService.addUserToGroup(userService.find(user_id).get(), groupService.find(group_id).get(), user);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    // TO DO: Authorization based on user login implementation
-    @PreAuthorize("hasRole('ROLE_PREMIMUM') && user_id == groupService.find(group_id).get().getGroupCreator()")
-    @PutMapping(value = "/{id}/groups/{id_group}/remove_user", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('PREMIUM') && (#user_id == authentication.principal.user.id)")
+    @PutMapping(value = "/{user_id}/groups/{group_id}/remove_user", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> removeUserFromGroup(@PathVariable Integer user_id, @PathVariable Integer group_id, @RequestBody User user){
         Objects.requireNonNull(user);
         if (!groupService.exists(group_id) || userService.find(user_id).isEmpty()){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        userService.removeUserFromGroup(userService.find(user_id), groupService.find(group_id), user);
+        if (!Objects.equals(user_id, groupService.find(group_id).get().getGroupCreator().getId())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        if (!groupService.find(group_id).get().getUsers().contains(user)){
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
+        userService.removeUserFromGroup(userService.find(user_id).get(), groupService.find(group_id).get(), user);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    // TO DO: Authorization based on user login implementation
-    @PreAuthorize("(!#groupService.find(id_group).get().groupCreator() == id)")
-    @DeleteMapping(value = "/{id}/groups/{id_group}")
-    public ResponseEntity<Void> deleteGroup(@PathVariable Integer id, @PathVariable Integer id_group){
-        return groupService.find(id_group)
-                .map(group -> {
-                    groupService.remove(group);
-                    return new ResponseEntity<Void>(ResponseEntity.ok().build().getStatusCode());
-                })
-                .orElseGet(() -> ResponseEntity.notFound().build());
-        // Not sure if the simplified version isn't worse to read and understand...
-        /*
-        Optional<Group> group = groupService.find(id_group);
+    @PreAuthorize("hasRole('PREMIUM') && (#user_id == authentication.principal.user.id)")
+    @DeleteMapping(value = "/{user_id}/groups/{group_id}")
+    public ResponseEntity<Void> deleteGroup(@PathVariable Integer user_id, @PathVariable Integer group_id){
+        Optional<Group> group = groupService.find(group_id);
         if (group.isPresent()){
+            if (!Objects.equals(user_id, group.get().getGroupCreator().getId())){
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
             groupService.remove(group.get());
             return new ResponseEntity<>(ResponseEntity.ok().build().getStatusCode());
         } else{
             return new ResponseEntity<>(ResponseEntity.notFound().build().getStatusCode());
         }
-         */
     }
 }
